Bug fix: Edit button not working
The edit form in the accordian element was not conditionally rendering
The condition to conditionally render was originally a loose check between
data types (id on the client is a string and from the data base a number).
The loose check got changed to strict check which was preventing the rerender.
Changed the id check back to loose check